requirements.txt
appdirs==1.4.4
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==2.9.2
distlib==0.3.1
Django==3.1.1
djangorestframework==3.11.1
dynaconf==3.1.0
filelock==3.0.12
hvac==0.10.5
idna==2.10
importlib-metadata==1.7.0
knox==0.1.0
loguru==0.5.1
Markdown==3.2.2
pycparser==2.20
pytz==2020.1
requests==2.24.0
shortuuid==1.0.1
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.10
virtualenv==20.0.31
win32-setctime==1.0.1
zipp==3.1.0
